# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UsuarioDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"perfil": {
						"type": "string",
						"enum": [
							"admin",
							"usuario"
						]
					}
				},
				"required": [
					"id",
					"nome",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequestDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"senha"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Usuario.Exclude_keyofUsuario.id__": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"senha": {
						"type": "string"
					},
					"perfil": {
						"type": "string",
						"enum": [
							"admin",
							"usuario"
						]
					}
				},
				"required": [
					"nome",
					"email",
					"senha",
					"perfil"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Usuario.id_": {
				"$ref": "#/components/schemas/Pick_Usuario.Exclude_keyofUsuario.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"BasicResponseDto": {
				"properties": {
					"mensagem": {
						"type": "string"
					},
					"objeto": {}
				},
				"required": [
					"mensagem",
					"objeto"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Omit_UsuarioDto.id__": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"perfil": {
						"type": "string",
						"enum": [
							"admin",
							"usuario"
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
		"/usuarios/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"usuario": {
											"$ref": "#/components/schemas/UsuarioDto"
										},
										"token": {
											"type": "string"
										}
									},
									"required": [
										"usuario",
										"token"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensagem": {
											"type": "string"
										}
									},
									"required": [
										"mensagem"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensagem": {
											"type": "string"
										}
									},
									"required": [
										"mensagem"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequestDto"
							}
						}
					}
				}
			}
		},
		"/usuarios": {
			"post": {
				"operationId": "CadastrarUsuario",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsuarioDto"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Usuario.id_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListarTodosUsuarios",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UsuarioDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Lista todos os usuários cadastrados.",
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": []
			}
		},
		"/usuarios/{id}": {
			"get": {
				"operationId": "FiltrarUsuarioPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsuarioDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensagem": {
											"type": "string"
										}
									},
									"required": [
										"mensagem"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Busca um usuário específico pelo seu ID.",
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [
					{
						"description": "O identificador numérico do usuário. Ex: 123",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletarUsuario",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensagem": {
											"type": "string"
										}
									},
									"required": [
										"mensagem"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deleta um usuário específico pelo seu ID.",
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "AtualizarUsuario",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsuarioDto"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"mensagem": {
											"type": "string"
										}
									},
									"required": [
										"mensagem"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Atualiza os dados de um usuário existente.",
				"tags": [
					"Usuario"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Omit_UsuarioDto.id__"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3040/api",
			"description": "Servidor de Desenvolvimento"
		}
	]
}